requirements.txt
asn1crypto==0.24.0
blinker==1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.1
facebook-sdk==3.1.0
Flask==1.0.2
Flask-Mail==0.9.1
Flask-Script==2.0.6
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pycparser==2.19
PyJWT==1.6.4
PyMySQL==0.9.2
requests==2.20.1
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
